Fixed site re-ordering bug
If sites appear in the final state in a different order than in the initial state the term get reordered first to match the order of the initial state
However if the same site name appears more than one time on each side they are assumed to be in the corrected order. This convention is necessary for describing diffusion processes.